refactor and comments
